# [PERF] web: improve perf of `search_panel_select_multi_range`