@robodoo 2024-07-02
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/171684.png)](https://mergebot.odoo.com/odoo/odoo/pull/171684)