@adda-odoo 2024-07-16
Closing because changes relevant to this PR were made here https://github.com/odoo/odoo/pull/173104